functions/requirements.txt
google-cloud-compute
google-cloud-monitoring
pythonping